refactor<BE> : 리그 Api 요청 방식 추가 및 게시글 내림차순 정렬 추가 및 ApiId 자료형 Long으로 변경

See merge request s07-webmobile1-sub2/S07P12B204!90

+++ backend/src/main/java/com/ssafy/cheertogether/league/domain/League.java
@@ -28,6 +28,6 @@ public class League {
 
 	private String name;
 	private String logo;
-	private int apiId;
+	private Long apiId;
 	private String hanName;
 }
